Enable mUTF-7/UTF-8 depending on server capability
I don't like how this got, but at least it works. The best way would be to
patch flexi-streams with support for mUTF-7 external-format, but I've looked
into it and it seems like a lot of work. For now this will do, although it's
probably slower and conses more than it should.